+++ ptui/src/History.tsx
@@ -1,7 +1,8 @@
-import * as I from 'immutable';
+import I from 'immutable';
 import * as React from "react";
 
 import * as M from './Model';
+import * as A from './Actions';
 import * as T from './PTTypes';
 
 
@@ -17,7 +18,7 @@ export function History(): JSX.Element {
             key={snapshot_index.toString() + "-" + log_index.toString()}>
             <GameLog log={log} creatures={creatures} />
             <button className="material-icons"
-              onClick={() => M.sendCommand({ t: "Rollback", snapshot_index, log_index })}
+              onClick={() => A.sendCommand({ t: "Rollback", snapshot_index, log_index })}
             >history</button>
           </div>)
     )
@@ -51,19 +52,19 @@ export function GameLog(props: { log: T.GameLog; creatures: I.Map<T.CreatureID, 
         <div>Success? {log.success.toString()}</div>
       </div>;
     case "CreateFolder":
-      return <div><div>Created Folder</div><div>{M.folderPathToString(log.path)}</div></div>;
+      return <div><div>Created Folder</div><div>{T.folderPathToString(log.path)}</div></div>;
     case "RenameFolder":
       return <div>Renamed Folder</div>;
     case "DeleteFolderItem":
-      return <div>Deleted folder item in {M.folderPathToString(log.path)}</div>;
+      return <div>Deleted folder item in {T.folderPathToString(log.path)}</div>;
     case "MoveFolderItem":
-      return <div>Moved folder item from {M.folderPathToString(log.path)}
-        to {M.folderPathToString(log.newPath)}</div>;
+      return <div>Moved folder item from {T.folderPathToString(log.path)}
+        to {T.folderPathToString(log.newPath)}</div>;
     case "CopyFolderItem":
-      return <div>Copied folder item from {M.folderPathToString(log.source)}
-        to {M.folderPathToString(log.dest)}</div>;
+      return <div>Copied folder item from {T.folderPathToString(log.source)}
+        to {T.folderPathToString(log.dest)}</div>;
     case "CreateItem":
-      return <div>Created item {log.item.name} in {M.folderPathToString(log.path)}</div>;
+      return <div>Created item {log.item.name} in {T.folderPathToString(log.path)}</div>;
     case "EditItem":
       return <div>Edited item {log.item.name}</div>;
     case "CreateNote":
