+++ ptui/src/Main.tsx
@@ -1,20 +1,19 @@
 import * as React from "react";
 
 import * as M from "./Model";
+import {startPoll, RPI_URL} from "./Actions";
 
 export function Main(props: React.PropsWithChildren) {
   const fetchStatus = M.useState(s => s.fetchStatus);
   // kick off a fetch of the app
-  React.useEffect(() => {
-    M.startPoll();
-  });
+  React.useEffect(() => { startPoll(); }, []);
 
   if (fetchStatus === "Unfetched") {
     return <div>Waiting for initial data from server.</div>;
   } else if (fetchStatus === "Error") {
     return (
       <div>
-        There was an error fetching the application state from {M.RPI_URL}.
+        There was an error fetching the application state from {RPI_URL}.
         Trying again...
       </div>
     );
